+++ src/unit/enemy.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Default)]
 #[allow(unused)]
 pub struct Enemy {
-    pub speed: f32,
+    pub speed: f64,
 }
 
 impl Enemy {
-    pub fn with_speed(speed: f32) -> Self {
+    pub fn with_speed(speed: f64) -> Self {
         Self { speed }
     }
 }
