+++ src/unit/player_builder.rs
@@ -2,22 +2,22 @@ use super::player::Player;
 
 #[derive(Debug, Default)]
 pub struct PlayerBuilder {
-    health: u32,
-    speed: f32,
+    health: u8,
+    speed: f64,
 }
 
 impl PlayerBuilder {
-    #[allow(cippy::new_without_default)]
+    #[allow(clippy::new_without_default)]
     pub fn new() -> Self {
         Self::default()
     }
 
-    pub fn speed(mut self, speed: f32) -> Self {
+    pub fn speed(mut self, speed: f64) -> Self {
         self.speed = speed;
         self
     }
 
-    pub fn health(mut self, health: u32) -> Self {
+    pub fn health(mut self, health: u8) -> Self {
         self.health = health;
         self
     }
